Restructured constants
