+++ services/client/src/constants/index.ts
@@ -1,3 +1,4 @@
 export * from './addresses';
 export * from './ipfs';
 export * from './isProduction';
+export * from './api';
